interaktywna funckja, bez S

+++ f.hpp
@@ -1,6 +1,6 @@
 #ifndef F_HPP
 #define F_HPP
 extern "C" {
-	void f(unsigned char* arr, const int width, const int height, int S, int A, int B, int C, int D);
+	void f(unsigned char* arr, const int width, const int height, long long S, long long A, long long B, long long C, long long D);
 }
 #endif //F_HPP
